% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_config_file}
\alias{apply_config_file}
\title{Apply defaults arguments from YAML}
\usage{
apply_config_file(config_path, ...)
}
\description{
Load a YAML file and rerun the parent function using the arguments present in the YAML file.
}
\keyword{internal}
