% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_sequence}
\alias{check_sequence}
\title{Check sequences}
\usage{
check_sequence(
  database_path = NULL,
  release_path = NULL,
  interactive = TRUE,
  auto_update = FALSE,
  ...
)
}
\arguments{
\item{database_path}{(\code{character} of length 1) The file path or URL to the
spreadsheet with the sequence and taxonomy data, formatted as described in
\href{database_format.html}{the database format guide}. Default: \code{NULL}, prompt
user if needed.}

\item{release_path}{(\code{character} of length 1) The file path or URL to the
git/github repository the releases are stored in. Default: \code{NULL}, prompt
user if needed.}

\item{interactive}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to prompt the
user. This function is designed to be interactive, so \code{TRUE} is
recommended. Default: \code{TRUE}.}

\item{auto_update}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to allow
automated edits to the source database specified by \code{database_path}. Before
any edits are made, a backup of the file will be made in the same location
prefixed with ".backup_". Default: \code{FALSE}.}
}
\description{
Checks sequences for invalid characters and warn about sequence changes between versions.
}
\keyword{internal}
