% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\docType{data}
\name{config}
\alias{config}
\title{Default parameters}
\format{An object of class \code{list} of length 8.}
\usage{
config
}
\description{
Used as defaults for most functions in this package, to simplify their use. These might need to
be changed if the location of the Google Drive assets change or the directory structure of the
website on the server is changed.
}
\keyword{datasets}
