% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_release_dir}
\alias{find_release_dir}
\title{Find the folder with the release files}
\usage{
find_release_dir(path = getwd())
}
\arguments{
\item{path}{A path to a folder with a oomycetedbtools config file.}
}
\description{
Find the folder with the release files
}
