% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ncbi.R
\name{get_isolate_seqs}
\alias{get_isolate_seqs}
\title{Lookup gene sequences from NCBI}
\usage{
get_isolate_seqs(species, genes, isolates = NULL,
  extract_features = FALSE, gene_name_in_feature = TRUE,
  flanking = c(0, 0), db = "nucleotide", pause = 0.5, ...)
}
\arguments{
\item{species}{The names of species to look up.}

\item{genes}{The names of the genes to look up.}

\item{isolates}{The names of isolates to look up. Must be the same length as \code{species} if
used.}

\item{extract_features}{If TRUE, return the sequence for each feature in the sequence annotation
instead of the whole sequence.}

\item{gene_name_in_feature}{If TRUE, only return features that have one of the gene names
somewhere in their description. Only has an effect if extract_features is TRUE.}

\item{flanking}{A vector of length 2. The number of base pairs before and after the target gene
to include in the sequence returned. If the flanking sequence is not available, the sequence
will be considered incomplete.}

\item{db}{The name of the NCBI database to query. Only tested with "nucleotide", but a few others
might work.}

\item{pause}{The number of seconds to pause between each query. This avoids annoying NCBI and
having them block you IP address. Should be at least 0.35 seconds if you dont have an NCBI API
key and at least 0.1 seconds if you do.}

\item{...}{Additional terms to add to the search request for each species/isolate, see NCBI
documentation for a complete list:
http://www.ncbi.nlm.nih.gov/books/NBK25499/#\emph{chapter4_ESearch}}
}
\value{
A tibble (a type of data.frame) with the following columns, depending on settings:
\itemize{
\item species: The species search term associated with the result
\item isolate: The isolate search term associated with the result
\item query: The query used to search for sequences on NCBI
\item acc: The Genbank accession number
\item start: The index of the first base pair of the target gene in the original sequence
\item end: The index of the last base pair of the target gene in the original sequence
\item feature: The type of locus
\item type: The type of annotation for the sequence
\item name: The name of the locus
\item complete: If the locus was complete in the original sequence, according to the annotation
\item seq: The whole sequence the gene was found in
\item header: The name of the overall sequence the gene was found in
\item length: The length of the whole sequence
\item flank_start: The index of the first base pair of the target gene plus the flanking region in the original sequence
\item flank_end: The index of the last base pair of the target gene plus the flanking region in the original sequence
\item flank_complete: If the locus plus flanking region was complete in the original sequence
\item gene_seq: The target gene sequence plus the flanking region
\item gene_length: The length of the target gene plus the flanking region
}
}
\description{
Look for sequences of a particular gene for a list of species/isolates/genes from the Genbank
nucleotide database.
}
\examples{
\dontrun{

# Search for the whole seqeunces for with P. infestans Cox I
get_isolate_seqs(species = c("Phytophthora infestans"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 retmax = 100)

# Search for the just the gene sequence for  P. infestans Cox I
get_isolate_seqs(species = c("Phytophthora infestans"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 retmax = 100,
                 extract_features = TRUE)

# Search for all the gene sequences in whole sequences that contain P. infestans Cox I
get_isolate_seqs(species = c("Phytophthora infestans"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 retmax = 100,
                 extract_features = TRUE,
                 gene_name_in_feature = FALSE)

# Search for whole sequences for P. infestans Cox I for just some isolates
get_isolate_seqs(species = c("Phytophthora infestans", "Phytophthora infestans", "Phytophthora infestans"),
                 isolates = c("44", "580", "180"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"))

# Search for just the gene sequences for P. infestans Cox I for just some isolates
get_isolate_seqs(species = c("Phytophthora infestans", "Phytophthora infestans", "Phytophthora infestans"),
                 isolates = c("44", "580", "180"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 extract_features = TRUE)

# Search for P infestans ITS with flanking regions and subset for complete results
result <- get_isolate_seqs(species = c("Phytophthora"),
                           genes = c("internal transcribed spacer"),
                           retmax = 300,
                           extract_features = TRUE,
                           flanking = c(300, 300))
result[result$complete, ]
result[result$flank_complete, ]

}

}
