% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{get_last_release}
\alias{get_last_release}
\title{Get path to latest release}
\usage{
get_last_release(release_history_path)
}
\arguments{
\item{path}{A file path to either the release history CSV or a folder with a oomycetedbtools config file.}
}
\description{
Get the path to the current release of the database
}
\keyword{internal}
