% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_website.R
\name{get_release_data}
\alias{get_release_data}
\title{Get releases spreadsheet}
\usage{

  get_release_data(release_spreadsheet_path = config$release_spreadsheet_path)
}
\arguments{
\item{release_spreadsheet_path}{The path (including the filename) to the version of the release
spreadsheet on the server releative to the root of the website directory.}
}
\value{
tibble
}
\description{
Returns the releases spread sheet stored on the server.
}
