% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{has_internet}
\alias{has_internet}
\title{Taken from https://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r}
\usage{
has_internet()
}
\description{
Taken from https://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r
}
