% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{init_database}
\alias{init_database}
\title{Make first version of the datbase}
\usage{
init_database(output_path = "initial_release.fa",
  source_db = "https://docs.google.com/spreadsheets/d/1Waua9iUPWLRqkDE0KkgAANFVEXas-8I6nCgSF8bG_Mw")
}
\arguments{
\item{output_path}{Where to save the FASTA file output}

\item{source_db}{The URL to the google drive folder for oomycetedb}
}
\description{
Make first version of the database using the spreadsheet with sequences made by Val Feiland.
This was only use once to initialize the database.
Since then, the FASTA file has been edited directly
}
