% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_website.R
\name{make_blast_database}
\alias{make_blast_database}
\title{Make a BLAST database}
\usage{
make_blast_database(fasta_path, out_dir_path)
}
\arguments{
\item{fasta_path}{The path to the FASTA file to use to make the database from.}

\item{out_dir_path}{The path to the directory to store the BLAST database in.}
}
\value{
\code{NULL}
}
\description{
Make a BLAST database in the specificed location from a FASTA file.
The name of the BLAST database is made using the name of the input file.
}
