% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print_error_report}
\alias{print_error_report}
\title{Print an error message}
\usage{
print_error_report(error_msg, error_key = NULL)
}
\arguments{
\item{error_msg}{The message to print.}

\item{error_key}{If provided, this is used to lookup what to print based on the \code{error_msg}.}
}
\description{
Print a simplified description of an error message.
}
\keyword{internal}
