% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release}
\alias{release}
\title{Release a version of the database to the OomyceteDB}
\usage{
release(
  database_path = NULL,
  release_path = NULL,
  interactive = TRUE,
  validate = TRUE,
  auto_update = FALSE,
  overwrite = FALSE,
  description = NULL,
  google_user = NULL,
  google_pass = NULL,
  github_user = NULL,
  github_pass = NULL,
  config_path = find_config("oomycetedbtools_config.yml")
)
}
\arguments{
\item{database_path}{(\code{character} of length 1) The file path or URL to the
spreadsheet with the sequence and taxonomy data, formatted as described in
\href{database_format.html}{the database format guide}. Default: \code{NULL}, prompt
user if needed.}

\item{release_path}{(\code{character} of length 1) The file path or URL to the
git/github repository the releases are stored in. Default: \code{NULL}, prompt
user if needed.}

\item{interactive}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to prompt the
user. This function is designed to be interactive, so \code{TRUE} is
recommended. Default: \code{TRUE}.}

\item{validate}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to validate the
database source before attempting to create a release. It is highly
recommended to validate before a release. Default: TRUE.}

\item{auto_update}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to allow
automated edits to the source database specified by \code{database_path}. Before
any edits are made, a backup of the file will be made in the same location
prefixed with ".backup_". Default: \code{FALSE}.}

\item{overwrite}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to replace the
previous release with this release rather than creating a new one. Default:
\code{FALSE}.}

\item{description}{(\code{character} of length 1) A description of this release.
This will be shown publicly with the release. Default: prompt user.}

\item{google_user}{(\code{character} of length 1) A Google Drive user ID with
read/write permissions for the database folder. Required if \code{database_path}
is a URL to a Google spreadsheet. Default: \code{NULL}, prompt user if needed.}

\item{google_pass}{(\code{character} of length 1) The password associated with the
\code{google_user} option.  Required if \code{database_path} is a URL to a Google
spreadsheet. Default: \code{NULL}, prompt user if needed.}

\item{github_user}{(\code{character} of length 1) A Github user ID with read/write
permissions for the \code{oomycetedbdata} repository. Required if \code{release_path}
is a URL to a Github repository. Default: \code{NULL}, prompt user if needed.}

\item{github_pass}{(\code{character} of length 1) The password associated with the
\code{github_user} option. Required if \code{release_path} is a URL to a Github
repository. Default: \code{NULL}, prompt user if needed.}

\item{config_path}{(\code{character} of length 1) The file path to a configuration
file with defaults for these parameters.}
}
\description{
This is an interactive function that walks the user through the steps needed
to add a release of OomyceteDB to the \code{oomycetedbdata} package. The
\code{oomycetedbdata} package is the source of the data for the OomyceteDB
website, so this will effectively add a release to the website. This function
will do the following things:
\itemize{
\item Validate that the database is formatted correctly, as described in
\href{https://grunwaldlab.github.io/oomycetedbtools/articles/database_format.html}{the database format guide}.
\item Show the version of the database to be released and any changes to the user
for a manual check.
\item Upload the release to Github or store/modify a
version of the \code{oomycetedbdata} package on the local computer for manual
upload.
}
}
