% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{run_check}
\alias{run_check}
\title{Run a check function}
\usage{
run_check(check_func, check_desc, ...)
}
\arguments{
\item{check_func}{Function with code for check}

\item{check_desc}{The description of the check shown to the user while it runs}

\item{...}{Passed to check \code{check_func}}
}
\description{
Run a check function with a progress indicator and handle output for failed checks.
}
\keyword{internal}
