% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_website.R
\name{update_releases_spreadsheet}
\alias{update_releases_spreadsheet}
\title{Update release spreadsheet}
\usage{
update_releases_spreadsheet(website_root = config$website_root,
  googledrive_root_url = config$googledrive_root_url,
  release_spreadsheet_path = config$release_spreadsheet_path,
  release_spreadsheet_name = config$release_spreadsheet_name)
}
\arguments{
\item{website_root}{The absolute path to the root folder of the website}

\item{googledrive_root_url}{The url to the Google Drive folder containing database resoureces.}

\item{release_spreadsheet_path}{The path (including the filename) to the version of the release
spreadsheet on the server releative to the root of the website directory.}

\item{release_spreadsheet_name}{The name of the release spreadsheet on Google Drive with no
extension.}
}
\value{
\code{NULL}
}
\description{
Replaces the local (on server) releases spreadsheet with the one on Google Drive.
}
