% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_website.R
\name{update_website}
\alias{update_website}
\title{Update website}
\usage{
update_website(website_root = config$website_root,
  googledrive_root_url = config$googledrive_root_url,
  release_spreadsheet_path = config$release_spreadsheet_path,
  release_spreadsheet_name = config$release_spreadsheet_name,
  local_release_dir = config$local_release_dir,
  release_name_prefix = config$release_name_prefix,
  blast_database_dir = config$blast_database_dir,
  blast_path = config$blast_path)
}
\arguments{
\item{website_root}{The absolute path to the root folder of the website}

\item{googledrive_root_url}{The url to the Google Drive folder containing database resoureces.}

\item{release_spreadsheet_path}{The path (including the filename) to the version of the release
spreadsheet on the server releative to the root of the website directory.}

\item{release_spreadsheet_name}{The name of the release spreadsheet on Google Drive with no
extension.}

\item{local_release_dir}{The path to the directoy containing the archived FASTA files for all
releases on the server relative to the root of the website.}

\item{release_name_prefix}{Text added to the begining of the archived FASTA files on the server.}

\item{blast_database_dir}{The path to the directoy containing the archived blast datbases for all
releases on the server relative to the root of the website.}

\item{blast_path}{The path to the folder with blast executables}
}
\description{
Syncs the website with the information contained in the Google Drive folder.
This is run by a \href{https://en.wikipedia.org/wiki/Cron}{cron job} on the web
server every day.
}
\details{
It does the following:
\itemize{
\item Syncs the local (on server) releases spreadsheet with the Google Drive
releases spreadsheet
\item Looks for releases numbers that are in the spreadsheet, but have not yet
been downloaded
\item For each new release, the FASTA file for the release (specified by the file name in the
spreasheet) is downloaded to the server and a new BLAST database is made for it
}

The defaults for all the parameters are stored in the \code{config} list, defined in
"R/configuration.R". Typically, no parameters should need to be used.
}
