% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate the source database format}
\usage{
validate(
  database_path = NULL,
  release_path = NULL,
  interactive = TRUE,
  auto_update = FALSE,
  stop_on_fail = FALSE,
  verbose = TRUE,
  google_user = NULL,
  google_pass = NULL,
  github_user = NULL,
  github_pass = NULL,
  config_path = find_config("oomycetedbtools_config.yml")
)
}
\arguments{
\item{database_path}{(\code{character} of length 1) The file path or URL to the
spreadsheet with the sequence and taxonomy data, formatted as described in
\href{database_format.html}{the database format guide}. Default: \code{NULL}, prompt
user if needed.}

\item{release_path}{(\code{character} of length 1) The file path or URL to the
git/github repository the releases are stored in. Default: \code{NULL}, prompt
user if needed.}

\item{interactive}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to prompt the
user. This function is designed to be interactive, so \code{TRUE} is
recommended. Default: \code{TRUE}.}

\item{auto_update}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to allow
automated edits to the source database specified by \code{database_path}. Before
any edits are made, a backup of the file will be made in the same location
prefixed with ".backup_". Default: \code{FALSE}.}

\item{stop_on_fail}{(\code{TRUE}/\code{FALSE} of length 1) Whether or not to stop the
validation on the fist failed check. Even when \code{FALSE}, some checks will
stop the validation when they fail. Default: \code{FALSE}.}

\item{verbose}{(\code{TRUE}/\code{FALSE} of length 1) Print extra information.}

\item{google_user}{(\code{character} of length 1) A Google Drive user ID with
read/write permissions for the database folder. Required if \code{database_path}
is a URL to a Google spreadsheet. Default: \code{NULL}, prompt user if needed.}

\item{google_pass}{(\code{character} of length 1) The password associated with the
\code{google_user} option.  Required if \code{database_path} is a URL to a Google
spreadsheet. Default: \code{NULL}, prompt user if needed.}

\item{github_user}{(\code{character} of length 1) A Github user ID with read/write
permissions for the \code{oomycetedbdata} repository. Required if \code{release_path}
is a URL to a Github repository. Default: \code{NULL}, prompt user if needed.}

\item{github_pass}{(\code{character} of length 1) The password associated with the
\code{github_user} option. Required if \code{release_path} is a URL to a Github
repository. Default: \code{NULL}, prompt user if needed.}

\item{config_path}{(\code{character} of length 1) The file path to a configuration
file with defaults for these parameters.}
}
\description{
Check that the database spreadsheet does not have any issues that would cause
problems when hosting the website.
}
