% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{get_release_data}
\alias{get_release_data}
\title{Get release metadata table}
\usage{
get_release_data(path = getwd())
}
\arguments{
\item{path}{A file path to either the release history CSV or a folder with a oomycetedbtools config file.}
}
\description{
Return the parsed table of data on the releases.
}
